[Ide]Stop tooltip from appearing after completion list has closed
Commit 176ab0480dd24fa2e3623ea5b85c228dd4036ab1 introduced a bug where the
tooltip window would appear and stay open after the completion list was closed.

There is a chance that the hack using a timeout to change opacity gets run
after the tooltip window should have been closed. This can lead to a timing
issue, where the tooltip window appears and stays open  after the completion
list has closed.

Checking if the parent is visible before changing the opacity stops the tooltip
window from appearing when it shouldn't.